+++ src/Repository/InvoiceConfigurationRepositoryDecoration.php
@@ -1,0 +1,44 @@
+<?php
+
+/**
+ * Copyright © MB Arbatos Klubas. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace FreshAdvance\Invoice\Repository;
+
+use FreshAdvance\Invoice\DataType\InvoiceConfiguration;
+use FreshAdvance\Invoice\DataType\InvoiceConfigurationInterface;
+use FreshAdvance\Invoice\Exception\InvoiceConfigurationNotFound;
+use FreshAdvance\Invoice\Settings\ModuleSettingsInterface;
+
+class InvoiceConfigurationRepositoryDecoration implements InvoiceConfigurationRepositoryInterface
+{
+    public function __construct(
+        private InvoiceConfigurationRepositoryInterface $originalRepository,
+        private ModuleSettingsInterface $moduleSettings,
+    ) {
+    }
+
+    public function getByOrderId(string $orderId): InvoiceConfigurationInterface
+    {
+        try {
+            $result = $this->originalRepository->getByOrderId($orderId);
+        } catch (InvoiceConfigurationNotFound) {
+            $result = new InvoiceConfiguration(
+                orderId: $orderId,
+                date: $this->moduleSettings->getInvoiceDateFormat(),
+                number: $this->moduleSettings->getInvoiceNumberFormat(),
+            );
+        }
+
+        return $result;
+    }
+
+    public function save(InvoiceConfigurationInterface $invoiceConfiguration): void
+    {
+        $this->originalRepository->save($invoiceConfiguration);
+    }
+}