+++ src/Repository/InvoiceConfigurationRepositoryInterface.php
@@ -13,7 +13,7 @@ use FreshAdvance\Invoice\DataType\InvoiceConfigurationInterface;
 
 interface InvoiceConfigurationRepositoryInterface
 {
-    public function getByOrderId(string $orderId): ?InvoiceConfigurationInterface;
+    public function getByOrderId(string $orderId): InvoiceConfigurationInterface;
 
     public function save(InvoiceConfigurationInterface $invoiceConfiguration): void;
 }